postsurvey: Postsurvey responses

Data from the postsurvey responses of the course participants in DS4OWD

Usage

postsurvey

Format

A tibble with 25 rows and 18 variables

course_completion

Whether the course was completed by the respondent or not

rating_overall

Overall course rating

expectations

Expectations from the course

learning_objectives

Whether learning objectives were met

rating_structure

Rating for the structure of the course

rating_ins_knowledge

Rating for the instructor's knowledge of the course material

rating_ins_clarity

Rating for the instructor's clarity

rating_ins_comp

Rating for the instructor's competency

rating_ins_app

Rating for approachability of the instructor

feedback

Feedback for the course

rating_self_r_comp

Self rating for R competency

rating_self_vc_comp

Self rating for version control competency

conf_skill_app

Confidence in applying skills learned during the course

most_liked

Most liked aspects of the course

least_liked

Least liked aspects of the course

suggestion

Suggestions for the course

other_comments

Any other comments

reason_non_participation

If course was not completed