presurvey: Responses from the pre-course survey

Data from the pre-course survey, including previous programming experience, ease of programming, data format, current level of LLM usage, organisation type, and barriers to learning.

Usage

presurvey

Format

A tibble with 212 rows and 15 variables

id

ID of survey response

google_timestamp

zoom_timestamp

prog_exp

Previous programming experience

prog_exp_r

Previous programming experience in R

prog_ease_max_list

Ease of programming

data_format

Format of data

current_use_llm

Current level of LLM usage

work_organisation

Organisation type where participant works

barrier_time_availability

If time availability is a barrier to learning

barrier_supervisor_interests

If supervisor's interest is a barrier to learning

barrier_internet_connection

If internet connection is a barrier to learning

barrier_electricity

If electricity availability is a barrier to learning

barrier_computer

If availability of a compiuter is a barrier to learning

interest_share_dataset

If interested in sharing a dataset